Update package name
