+++ src/main/java/com/mapzen/pelias/gson/Properties.java
@@ -1,8 +1,9 @@
-package com.mapzen.android.gson;
+package com.mapzen.pelias.gson;
 
-import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+
+import javax.annotation.Generated;
 
 @Generated("org.jsonschema2pojo")
 public class Properties {
